moved tests to plotting
